third_party/WebKit/LayoutTests/fast/dom/DeviceOrientation/basic-operation-absolute-expected.txt

Tests basic operation of deviceorientationabsolute event using mock data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.alpha is mockAlpha
PASS event.beta is mockBeta
PASS event.gamma is mockGamma
PASS event.absolute is mockAbsolute
PASS successfullyParsed is true

TEST COMPLETE
